air_hockey_agent/requirements.txt
numpy>=1.21.0
torch>=1.9.0
mushroom-rl>=1.7.0
pyyaml>=5.4.1
air-hockey-challenge>=1.0.0 